🚚 Move: 통계 페이지 이동
