+++ src/pages/statistics/WeeklyStatisticsPage.tsx
