+++ src/router.tsx
@@ -12,8 +12,8 @@ import GroupCreatePage from '@/pages/GroupCreatePage';
 import GroupSelectPage from '@/pages/GroupSelectPage';
 import GroupSettingPage from '@/pages/GroupSettingPage';
 import HomePage from '@/pages/home/HomePage';
-import WeeklyStatisticsPage from '@/pages/WeeklyStatisticsPage';
-import MonthlyStatisticsPage from '@/pages/MonthlyStatisticsPage';
+import WeeklyStatisticsPage from '@/pages/statistics/WeeklyStatisticsPage';
+import MonthlyStatisticsPage from '@/pages/statistics/MonthlyStatisticsPage';
 import PresetSettingPage from '@/pages/PresetSettingPage';
 import MyPage from '@/pages/MyPage';
 import MyPageEditPage from '@/pages/MyPageEditPage';
